# Conflicts:
#	src/gumbo.h
#	src/parser.c
#	src/tag.c
#	src/tag_gperf.h
#	src/tokenizer.c

+++ src/error.c
@@ -37,7 +37,7 @@ static int PRINTF(2) print_message (
   ...
 ) {
   va_list args;
-  int remaining_capacity = output->capacity - output->length;
+  size_t remaining_capacity = output->capacity - output->length;
   va_start(args, format);
   int bytes_written = vsnprintf (
     output->data + output->length,
